vermouth/Index/view.py
# -*- coding: utf-8 -*-
from __future__ import absolute_import, unicode_literals
from django.contrib.auth.decorators import login_required
from django.shortcuts import render_to_response
from django.template import RequestContext


def index(request):
    return render_to_response('douban/index.html',RequestContext(request))